+++ pages/index.js
@@ -1,6 +1,6 @@
 import MatchList from "../components/MatchList";
-import AppHeader from "../components/AppHeader/AppHeader";
-import AppFooter from "../components/AppFooter/AppFooter";
+import AppHeader from "../components/AppHeader/index.js";
+import AppFooter from "../components/AppFooter/index.js";
 
 export default function Home({matches}) {
   return (
